add a Buffer changelog
Have Buffer-editing commands return a changelog of what was changed,
so that callers (Window) can use this information to update
the Buffer's representation (ncurses window)

TODO: investigate implications of this. Currently exposes internals
by returning an iterator to the entire buffer. Perhaps instead make
a copy. Perhaps make it a vector of strings of all the lines that
were changed.

Also ignored the executable file.